# Backend/internal/interfaces/api/review_handler.go
package api

import (
	"github.com/gin-gonic/gin"

	"uniStore/Backend/internal/domain/services"
)

// ReviewHandler handles HTTP requests related to reviews
type ReviewHandler struct {
	reviewService services.ReviewService
}

// NewReviewHandler creates a new review handler
func NewReviewHandler(reviewService services.ReviewService) *ReviewHandler {
	return &ReviewHandler{
		reviewService: reviewService,
	}
}

// CreateReview creates a new review
// @Summary Create a new review
// @Description Creates a new review for a game
// @Tags Reviews
// @Accept json
// @Produce json
// @Param review body dto.ReviewCreateDTO true "Review data"
// @Success 201 {object} dto.ReviewCreateDTO "Review created successfully"
// @Failure 400 {object} map[string]interface{} "Invalid input"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 404 {object} map[string]interface{} "Game not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/reviews [post]
func (h *ReviewHandler) CreateReview(c *gin.Context) {
	// Implementation to be added
}

// GetReviewByID retrieves a review by ID
// @Summary Get a review by ID
// @Description Returns a review by its ID
// @Tags Reviews
// @Accept json
// @Produce json
// @Param review_id path int true "Review ID"
// @Success 200 {object} dto.ReviewResponseDTO "Review details"
// @Failure 400 {object} map[string]interface{} "Invalid review ID"
// @Failure 404 {object} map[string]interface{} "Review not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /api/v1/reviews/{review_id} [get]
func (h *ReviewHandler) GetReviewByID(c *gin.Context) {
	// Implementation to be added
}

// GetReviewsByGameID retrieves all reviews for a game
// @Summary Get reviews by game ID
// @Description Returns all reviews for a specific game
// @Tags Reviews
// @Accept json
// @Produce json
// @Param game_id path int true "Game ID"
// @Success 200 {array} dto.ReviewResponseDTO "List of reviews"
// @Failure 400 {object} map[string]interface{} "Invalid game ID"
// @Failure 404 {object} map[string]interface{} "Game not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /api/v1/reviews/game/{game_id} [get]
func (h *ReviewHandler) GetReviewsByGameID(c *gin.Context) {
	// Implementation to be added
}

// UpdateReview updates a review
// @Summary Update a review
// @Description Updates an existing review
// @Tags Reviews
// @Accept json
// @Produce json
// @Param review_id path int true "Review ID"
// @Param user_id path int true "User ID"
// @Param review body dto.ReviewUpdateDTO true "Updated review data"
// @Success 200 {object} dto.ReviewUpdateDTO "Review updated successfully"
// @Failure 400 {object} map[string]interface{} "Invalid input"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 403 {object} map[string]interface{} "Forbidden"
// @Failure 404 {object} map[string]interface{} "Review not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/reviews/{review_id}/user/{user_id} [patch]
func (h *ReviewHandler) UpdateReview(c *gin.Context) {
	// Implementation to be added
}

// DeleteReview deletes a review
// @Summary Delete a review
// @Description Deletes an existing review
// @Tags Reviews
// @Accept json
// @Produce json
// @Param review_id path int true "Review ID"
// @Param user_id path int true "User ID"
// @Success 200 {object} map[string]interface{} "Review deleted successfully"
// @Failure 400 {object} map[string]interface{} "Invalid input"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 403 {object} map[string]interface{} "Forbidden"
// @Failure 404 {object} map[string]interface{} "Review not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/reviews/{review_id}/user/{user_id} [delete]
func (h *ReviewHandler) DeleteReview(c *gin.Context) {
	// Implementation to be added
}